+++ plugins/source/azure/resources/plugin/plugin.go
@@ -10,6 +10,9 @@ import (
 )
 
 var (
+	Name    = "azure"
+	Kind    = "source"
+	Team    = "cloudquery"
 	Version = "development"
 )
 
@@ -82,9 +85,11 @@ func titleTransformer(table *schema.Table) error {
 
 func Plugin() *plugin.Plugin {
 	return plugin.NewPlugin(
-		"azure",
+		Name,
 		Version,
 		NewClient,
 		plugin.WithJSONSchema(spec.JSONSchema),
+		plugin.WithKind(Kind),
+		plugin.WithTeam(Team),
 	)
 }