+++ plugins/source/hubspot/resources/plugin/plugin.go
@@ -19,6 +19,9 @@ import (
 )
 
 var (
+	Name    = "hubspot"
+	Kind    = "source"
+	Team    = "cloudquery"
 	Version = "development"
 )
 
@@ -115,8 +118,10 @@ func getTables() schema.Tables {
 
 func Plugin() *plugin.Plugin {
 	return plugin.NewPlugin(
-		"cloudquery-hubspot",
+		Name,
 		Version,
 		newClient,
+		plugin.WithKind(Kind),
+		plugin.WithTeam(Team),
 	)
 }