+++ plugins/source/okta/resources/plugin/plugin.go
@@ -5,13 +5,18 @@ import (
 )
 
 var (
+	Name    = "okta"
+	Kind    = "source"
+	Team    = "cloudquery"
 	Version = "development"
 )
 
 func Plugin() *plugin.Plugin {
 	return plugin.NewPlugin(
-		"okta",
+		Name,
 		Version,
 		Configure,
+		plugin.WithKind(Kind),
+		plugin.WithTeam(Team),
 	)
 }