+++ cli/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to CloudQuery will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.25.0](https://github.com/cloudquery/cloudquery/compare/cli-v3.24.0...cli-v3.25.0) (2023-10-18)
+
+
+### Features
+
+* Addon CLI ([#14564](https://github.com/cloudquery/cloudquery/issues/14564)) ([33e7833](https://github.com/cloudquery/cloudquery/commit/33e7833875de8f1ec6c49e2366bf918a5bc59b91))
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.15.2 ([#14662](https://github.com/cloudquery/cloudquery/issues/14662)) ([e274fe4](https://github.com/cloudquery/cloudquery/commit/e274fe419f6cacdf62547cd7134f40916e5ddd96))
+
 ## [3.24.0](https://github.com/cloudquery/cloudquery/compare/cli-v3.23.1...cli-v3.24.0) (2023-10-18)